+deleted GlideApp.kt +exchanged annotationProcessor with kapt +moved HeaderInterceptor.kt to quotes/data/remote
