+++ app/build.gradle
@@ -17,7 +17,7 @@ android {
     defaultConfig {
         applicationId "com.example.dogsapp"
         minSdk 26
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -42,9 +42,9 @@ android {
 }
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.5.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'androidx.core:core-ktx:1.8.0'
+    implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.test.ext:junit-ktx:1.1.3'
 
@@ -52,14 +52,14 @@ dependencies {
     def room_version = "2.4.3"
 
     implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
 
     // To use Kotlin annotation processing tool (kapt)
     implementation 'androidx.room:room-ktx:2.4.3'
 
     kapt "androidx.room:room-compiler:$room_version"
 
-    def nav_version = "2.5.1"
+    def nav_version = "2.5.2"
 
     // Java language implementation
     implementation "androidx.navigation:navigation-fragment:$nav_version"
@@ -93,14 +93,14 @@ dependencies {
 
     //DaggerHilt
     implementation "com.google.dagger:hilt-android:2.40"
-    kapt "com.google.dagger:hilt-compiler:2.40"
+    kapt "com.google.dagger:hilt-compiler:2.40.1"
 
     //glide
     implementation 'com.github.bumptech.glide:glide:4.13.2'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.13.2'
+    kapt 'com.github.bumptech.glide:compiler:4.13.2'
 
     //Material 3
-    implementation 'com.google.android.material:material:1.5.0-alpha04'
+    implementation 'com.google.android.material:material:1.8.0-alpha01'
 
     //Lottie animation
     implementation "com.airbnb.android:lottie:5.2.0"
