--- app/src/main/java/com/example/dogsapp/dogs/ui/GlideApp.kt
@@ -1,8 +1,0 @@
-package com.example.dogsapp.dogs.ui
-
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
-
-@GlideModule
-class GlideApp: AppGlideModule() {
-}
